@@ -1,15 +1,15 @@
 package routes
 
 import (
-	"golang.org/x/crypto/bcrypt"
 	"encoding/json"
+	"golang.org/x/crypto/bcrypt"
 	"log"
 	"math"
+	"math/rand"
 	"net/http"
 	"strconv"
 	"strings"
 	"time"
-	"math/rand"
 
 	"msmf/database"
 	"msmf/utils"
@@ -55,65 +55,58 @@ func CreateReferral(w http.ResponseWriter, r *http.Request) {
 	tokenCookie, err := r.Cookie("token")
 	// Invalid token
 	if err != nil || !utils.ValidateToken(tokenCookie.Value) {
-		w.WriteHeader(http.StatusUnauthorized)
-		resp["error"] = "Unauthorized"
-	} else {
-		user := database.User{}
-		result := database.DB.Where("token = ?", tokenCookie.Value).First(&user)
-		// Error grabbing the user
-		if result.Error != nil {
-			w.WriteHeader(http.StatusUnauthorized)
-			resp["error"] = "Unauthorized"
+		utils.ErrorJSON(w, http.StatusUnauthorized, "Unauthorized")
+		return
+	}
+	user := database.User{}
+	result := database.DB.Where("token = ?", tokenCookie.Value).First(&user)
+	// Error grabbing the user
+	if result.Error != nil {
+		utils.ErrorJSON(w, http.StatusUnauthorized, "Unauthorized")
+		return
 		// Check user permissions
-		} else {
-			hasPerms := utils.CheckPermissions(&utils.PermCheck{
-				FKTable: "perms_per_users",
-				Perms: "invite_user",
-				PermTable: "user_perms",
-				Search: user.Token,
-				SearchCol: "token",
-				SearchTable: "users",
-			})
-			// User doesn't have correct perms
-			if !hasPerms {
-				w.WriteHeader(http.StatusUnauthorized)
-				resp["error"] = "Unauthorized"
-			// Create the invite code
-			} else {
-				// Make sure to seed the random
-				// Could use crypto random, but this doesn't really need to be secure
-				// It's not a secret, nor does it really matter
-				rand.Seed(time.Now().Unix())
+	}
+	hasPerms := utils.CheckPermissions(&utils.PermCheck{
+		FKTable:     "perms_per_users",
+		Perms:       "invite_user",
+		PermTable:   "user_perms",
+		Search:      user.Token,
+		SearchCol:   "token",
+		SearchTable: "users",
+	})
+	// User doesn't have correct perms
+	if !hasPerms {
+		utils.ErrorJSON(w, http.StatusUnauthorized, "Unauthorized")
+		return
+		// Create the invite code
+	}
+	// Make sure to seed the random
+	// Could use crypto random, but this doesn't really need to be secure
+	// It's not a secret, nor does it really matter
+	rand.Seed(time.Now().Unix())
 
-				// Loop until a valid code is found
-				for {
-					code := (rand.Int() % (int(math.Pow(10, 8)) - int(math.Pow(10,7)-1))) + int(math.Pow(10, 7))
-					referral := database.Referrer{
-						Code: code,
-						UserID: user.ID,
-						Expiration: time.Now().Add(time.Hour*6),
-					}
-					result = database.DB.Create(&referral)
-					if result.Error == nil {
-						resp["status"] = "Success"
-						resp["code"] = code
-						break
-					} else {
-						log.Println(result.Error.Error())
-					}
-				}
-			}
+	// Loop until a valid code is found
+	for {
+		code := (rand.Int() % (int(math.Pow(10, 8)) - int(math.Pow(10, 7)-1))) + int(math.Pow(10, 7))
+		referral := database.Referrer{
+			Code:       code,
+			UserID:     user.ID,
+			Expiration: time.Now().Add(time.Hour * 6),
+		}
+		result = database.DB.Create(&referral)
+		if result.Error == nil {
+			resp["status"] = "Success"
+			resp["code"] = code
+			break
+		} else {
+			log.Println(result.Error.Error())
 		}
 	}
 
-
 	// Write out the body
-	out, err := json.Marshal(resp)
-	if err != nil {
-		log.Println(err)
-	}
-	w.Write(out)
+	w.Write(utils.ToJSON(resp))
 }
+
 // Refer allows users with proper permissions to send someone an invite code
 // in order to make an account
 func Refer(w http.ResponseWriter, r *http.Request) {
@@ -121,13 +114,8 @@ func Refer(w http.ResponseWriter, r *http.Request) {
 	// Can't error due to regex checking on subrouter
 	code, _ := strconv.Atoi(parts[len(parts)-1])
 	resp := make(map[string]interface{})
-	if code < int(math.Pow(10, 7)) || code > int(math.Pow(10, 8)) - 1 {
-		resp["error"] = "Not Found"
-		out, err := json.Marshal(resp)
-		if err != nil {
-			log.Println(err)
-		}
-		w.Write(out)
+	if code < int(math.Pow(10, 7)) || code > int(math.Pow(10, 8))-1 {
+		utils.ErrorJSON(w, http.StatusNotFound, "Not Found")
 		return
 	}
 
@@ -135,64 +123,59 @@ func Refer(w http.ResponseWriter, r *http.Request) {
 	referrer := database.Referrer{}
 	result := database.DB.Preload("User").First(&referrer, code)
 	if result.Error != nil {
-		w.WriteHeader(http.StatusNotFound)
-		resp["error"] = "Not Found"
-	// Handle referral expiration
+		utils.ErrorJSON(w, http.StatusNotFound, "Not Found")
+		return
+		// Handle referral expiration
 	} else if referrer.Expiration.Before(time.Now()) {
 		// Remove the code from the db
 		database.DB.Delete(&referrer)
-		w.WriteHeader(http.StatusNotFound)
-		resp["error"] = "Not Found"
+		utils.ErrorJSON(w, http.StatusNotFound, "Not Found")
+		return
 	} else if r.Method == "GET" {
 		resp["username"] = referrer.User.Username
 		resp["expiration"] = referrer.Expiration
-	// Must be Method PUT
-	} else {
-		// Get JSON of body of PUT
-		body := make(map[string]string)
-		err := json.NewDecoder(r.Body).Decode(&body)
-		// Invalid data
-		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			resp["error"] = "Bad Request"
-		// Could decode body
-		} else {
-			// Check if any of the following are missing
-			if len(body["username"]) == 0 || len(body["password"]) == 0 {
-				w.WriteHeader(http.StatusBadRequest)
-				resp["error"] = "Bad Request"
-			} else {
-				// Generate the password hash
-				hash, err := bcrypt.GenerateFromPassword([]byte(body["password"]), bcrypt.DefaultCost)
-				// Something wrong with hashing
-				if err != nil {
-					w.WriteHeader(http.StatusBadRequest)
-					resp["error"] = err.Error()
-				} else {
-					// Attempt to make the user
-					result = database.DB.Create(&database.User{
-						Username: body["username"],
-						Password: hash,
-						ReferredBy: referrer.UserID,
-					})
-					// User already exists
-					if result.Error != nil {
-						w.WriteHeader(http.StatusUnauthorized)
-						resp["error"] = "user already exists"
-					} else {
-						// All good, remove one time referrer code
-						database.DB.Delete(&referrer)
-						resp["status"] = "Success"
-					}
-				}
-			}
-		}
+		// Must be Method PUT
 	}
 
-	// Finally write out the body
-	out, err := json.Marshal(resp)
+	// Get JSON of body of PUT
+	body := make(map[string]string)
+	err := json.NewDecoder(r.Body).Decode(&body)
+
+	// Invalid data
 	if err != nil {
-		log.Println(err)
+		utils.ErrorJSON(w, http.StatusBadRequest, "Bad Request")
+		return
 	}
-	w.Write(out)
+
+	// Check if any of the following are missing
+	if len(body["username"]) == 0 || len(body["password"]) == 0 {
+		utils.ErrorJSON(w, http.StatusBadRequest, "Bad Request")
+		return
+	}
+
+	// Generate the password hash
+	hash, err := bcrypt.GenerateFromPassword([]byte(body["password"]), bcrypt.DefaultCost)
+	// Something wrong with hashing
+	if err != nil {
+		utils.ErrorJSON(w, http.StatusBadRequest, err.Error())
+		return
+	}
+
+	// Attempt to make the user
+	result = database.DB.Create(&database.User{
+		Username:   body["username"],
+		Password:   hash,
+		ReferredBy: referrer.UserID,
+	})
+	// User already exists
+	if result.Error != nil {
+		utils.ErrorJSON(w, http.StatusBadRequest, err.Error())
+		return
+	}
+
+	// All good, remove one time referrer code
+	database.DB.Delete(&referrer)
+	resp["status"] = "Success"
+
+	w.Write(utils.ToJSON(resp))
 }
